+++ src/components/Input/index.ts
@@ -1,4 +1,4 @@
-// Input from copy-ui @ 2025-11-19.
+// Input from copy-ui @ 2025-12-07.
 
 export type { InputProps } from "./input";
 export { Input } from "./input";